backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-chess==1.999
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
pytest==8.0.1
black==24.1.1
isort==5.13.2
flake8==7.0.0 